Refactor: 코드 컨벤션
